+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/card_units.js
@@ -429,6 +429,7 @@ define([
       { id: "gwaio_cooldown_orbital", units: gwoGroup.orbitalFactories },
       { id: "gwaio_speed_structure", units: gwoGroup.nomadStructures },
       { id: "gwaio_enable_bounties" },
+      { id: "gwaio_enable_suddendeath" },
     ],
   };
 });