+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/cards.js
@@ -233,6 +233,7 @@ function gwoCard() {
         "gwaio_damage_titans",
         "gwaio_enable_bounties",
         "gwaio_enable_planetaryradar",
+        "gwaio_enable_suddendeath",
         "gwaio_health_titans",
         "gwaio_speed_structure",
         "gwaio_speed_titans",